# apis/filter.js
/**
 * Created by anahid on 7/30/17.
 */
var response = require("../common/const");

var mongojs = require('mongojs');
var db = mongojs('mongodb://localhost/SimpleServer');

function provide(router) {
    router.get('/filter/:model/:name/:value', function (req, res) {
        try {
            var query = {};
            query[req.params.name] = req.params.value;
            db.collection(req.params.model).find(query, function (err, docs) {
                if (err) {
                    console.log(err);
                    res.send(response.DB_ERROR);
                } else {
                    res.send({code: 0, result: docs});
                }
            });
        } catch (e) {
            console.log(response.GENERAL_ERROR)
        }
    });
}

exports.provide = provide;